+++ saltax/plugins/__init__.py
@@ -1,8 +1,8 @@
 from . import s_raw_records
 from .s_raw_records import *
 
-from . import s_records
-from .s_records import *
+from . import records
+from .records import *
 
-from . import s_peaklets
-from .s_peaklets import *
+from . import peaklets
+from .peaklets import *
